Updated twoDegreesOfSeparation function
Made findFollowers function in Graph.java more concise and added a person as an input to check that persons followers.

In socialNetwork.java I created a getFirstPerson function to return the first person from the input text for the twoDegreesOfSeparation function and built on the twoDegreesOfSeparation function. It now finds that first persons followers and stores them into a set. Then 2 for each loops go through and find the followers of the followers.