Task 5, created getMedianFollowers function
Created a function to get the median number of followers from all people in the graph. Includes a function to initialise the numbers of followers of each person into an array, and another function to find the median value and deal with the case if number of people is even. Task 5.