Started findBestMessagePropogater function
Created calculateReach function to start to find the best person to propogate a message. Built the function to work with a single person. Functions correctly after testing it. - Task 6

Added a reach integer to each person (node), will help calculate who is best for spreading a message.

Changed getNumEdges function to regular int from float. the integers are then converted to floats in the calculateDensity function in socialNetwork.java